Extract RPCRequestHandler
